Begin to refine matchmaking process
This is a first pass at trying to find the right domain terminology
for the internal participation in a matchmaking process.

I suspect that the Matchmaker needs a stronger hand in keeping track
of who is linked to who instead of deferring to the participants.

At the time of writing, you can test this out using the seed data by
issuing the following command:

```
StableMarriageMatchmaker.new(Person.mentees.includes(:preferences), Person.mentors.includes(:preferences)).run
```